refactor: 탭과 포스트카드, 함수 컴포넌트화 완료
